Extract common code from ReflektIrGenerationExtension and SmartReflektIrGenerationExtension to BaseReflektIrGenerationExtension
