+++ reflekt-plugin/src/main/kotlin/org/jetbrains/reflekt/plugin/generation/ir/SmartReflektIrGenerationExtension.kt
@@ -1,12 +1,9 @@
 package org.jetbrains.reflekt.plugin.generation.ir
 
-import org.jetbrains.reflekt.plugin.analysis.analyzer.IrInstancesAnalyzer
-
-import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
-import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
-
+import org.jetbrains.reflekt.plugin.analysis.analyzer.IrInstancesAnalyzer
+import org.jetbrains.reflekt.plugin.analysis.models.ir.IrInstances
 import java.io.File
 
 /**
@@ -17,21 +14,13 @@ import java.io.File
  * @param messageCollector
  */
 class SmartReflektIrGenerationExtension(
-    private val irInstancesAnalyzer: IrInstancesAnalyzer,
+    irInstancesAnalyzer: IrInstancesAnalyzer,
     private val classpath: List<File>,
     private val messageCollector: MessageCollector? = null,
-) : IrGenerationExtension {
-    /**
-     * Replaces IR in the SmartReflekt queries to the list of the found entities.
-     */
-    override fun generate(moduleFragment: IrModuleFragment, pluginContext: IrPluginContext) {
-        val irInstances = irInstancesAnalyzer.getIrInstances()
-        if (irInstances.isEmpty()) {
-            return
-        }
-        val storageClassGenerator = StorageClassGenerator(pluginContext)
-        val transformer = SmartReflektIrTransformer(irInstances, pluginContext, classpath, messageCollector, storageClassGenerator)
-        moduleFragment.transform(transformer, null)
-        storageClassGenerator.contributeInitializers(transformer.storageClasses)
-    }
+) : BaseReflektIrGenerationExtension(irInstancesAnalyzer) {
+    override fun getTransformer(
+        pluginContext: IrPluginContext,
+        irInstances: IrInstances,
+        storageClassGenerator: StorageClassGenerator
+    ): BaseReflektIrTransformer = SmartReflektIrTransformer(irInstances, pluginContext, classpath, messageCollector, storageClassGenerator)
 }
